Finish resolving merge conflicts
