+++ test/stat_tracker_test.rb
@@ -235,11 +235,11 @@ assert_equal 1, @stat_tracker.biggest_team_blowout("3")
 ##### Season Statistics Tests #####
 
   def test_biggest_bust
-    assert_equal [], @stat_tracker.biggest_bust("20122013")
+    assert_equal "Seattle Sounders FC", @stat_tracker.biggest_bust("20122013")
   end
 
   # def test_biggest_surprise
-  #   assert_equal [], @stat_tracker.biggest_surprise("20122013")
+  #   assert_equal [], @stat_tracker.biggest_surprise("20142015")
   # end
 
   def test_winningest_coach
